Move code in dedicated component
